fix(parser): adapt to tree-sitter 0.27 API changes
tree-sitter 0.27 turned QueryMatch::captures from a public field into an accessor method. Updated both call sites in the extractor and the cpp test helper to call m.captures(). Clippy also flagged i as u32 casts at every Node::child(i) indexing site — the new API accepts usize directly, so the casts are gone from comments.rs, cpp.rs, kotlin.rs and python.rs (6 sites).

+++ src/pipeline/parser/comments.rs
@@ -43,7 +43,7 @@ pub(crate) fn extract_comments(
                         // Try to find the doc_comment child node by iterating through children
                         let mut doc_text = None;
                         for i in 0..node.child_count() {
-                            if let Some(child) = node.child(i as u32)
+                            if let Some(child) = node.child(i)
                                 && child.kind() == "doc_comment"
                             {
                                 doc_text = Some(node_text(child, source));
